film part controllers

+++ src/main/java/sk/vaii/sem/semestralna_praca_vaii_backend/part_film/service/ActorService.java
@@ -2,7 +2,11 @@ package sk.vaii.sem.semestralna_praca_vaii_backend.part_film.service;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import sk.vaii.sem.semestralna_praca_vaii_backend.mapper.ActorMapper;
+import sk.vaii.sem.semestralna_praca_vaii_backend.part_film.dto.ActorAddDto;
+import sk.vaii.sem.semestralna_praca_vaii_backend.part_film.dto.ActorUpdateDto;
 import sk.vaii.sem.semestralna_praca_vaii_backend.part_film.entity.Actor;
+import sk.vaii.sem.semestralna_praca_vaii_backend.part_film.entity.FilmPartImage;
 import sk.vaii.sem.semestralna_praca_vaii_backend.part_film.repository.ActorRepository;
 
 import java.util.List;
@@ -12,12 +16,26 @@ import java.util.Optional;
 @AllArgsConstructor
 public class ActorService {
     private final ActorRepository actorRepository;
+    private final ActorMapper actorMapper;
+    private final FilmPartImageService filmPartImageService;
 
-    public Actor addActor(Actor actor) {
+    public Actor addActor(ActorAddDto actorAddDto) {
+        Actor actor = this.actorMapper.actorAddDtoToActor(actorAddDto);
+
+        //find image and set it ot actor
+        Optional<FilmPartImage> image = this.filmPartImageService.getImageById(actorAddDto.getImage_id());
+        image.ifPresent(actor::setActorImage);
+
         return this.actorRepository.save(actor);
     }
 
-    public Actor updateActor(Actor actor) {
+    public Actor updateActor(ActorUpdateDto actorUpdateDto) {
+        Actor actor = this.actorMapper.actorUpdateDtoToActor(actorUpdateDto);
+
+        //find image and set it ot actor
+        Optional<FilmPartImage> image = this.filmPartImageService.getImageById(actorUpdateDto.getImage_id().getId());
+        image.ifPresent(actor::setActorImage);
+
         return this.actorRepository.save(actor);
     }
 
@@ -34,7 +52,12 @@ public class ActorService {
     }
 
     public void deleteActor(Long id) {
-        this.actorRepository.deleteById(id);
+        Actor actor = this.actorRepository.getById(id);
+        if (actor.getFilms().isEmpty()) {
+            this.actorRepository.deleteById(id);
+        } else {
+            throw new IllegalStateException("Actor can't be assigned to any movie.");
+        }
     }
 
 }
